Fix bug in ASA calculations in dpocket
The get_surrounding_atoms_idx() function does a full pass through the PDB
and ignores hydrogen atoms but currently includes ligand atoms flagged by
dpocket. This leads to differing results for all subsequent ASA descriptors
(and the drug score).

The fix here just adds ligname as an argument back through the chain of
calls so that residues can be skipped by name. A null string is explicitly
passed for fpocket and mdpocket -- this shouldn't be a problem unless a
residue has an empty name in those programs and I assume that isn't allowed.

get_surrounding_atoms_idx() is also called inside writepocket.c, which is
shared by all three programs. The change here does NOT update ligname to
pass through, so the results of this call are probably still different
between fpocket and dpocket -- I'll regard this as aesthetic until I can
find descriptor issues.